temp_requirements.txt
certifi==2025.8.3
charset-normalizer==3.4.2
diskcache==5.6.3
filelock==3.16.1
fsspec==2025.3.0
hf-xet==1.1.5
huggingface-hub==0.34.3
idna==3.10
Jinja2==3.1.6
llama_cpp_python==0.3.14
MarkupSafe==2.1.5
numpy==1.24.4
packaging==25.0
PyYAML==6.0.2
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
tokenizers==0.20.3
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.13.2
urllib3==2.2.3
